feat: refactor and added new challenges
