+++ challenges/3/challenge.js
@@ -1,22 +1,37 @@
 /*
- * Calcular o MDC
+ * Requisição na API e componente de Loading
  */
 
 /* ENUNCIADO
  *
- * Você deve escrever uma função que calcula e retorna o MDC
- * (máximo divisor comum) entre dois números.
- * Para isso você pode usar as seguintes informações:
+ * Neste exercício você deverá escrever  uma função que faz
+ * uma requisição para nossa API. Em uma aplicação real,
+ * essa função seria chamada no início da renderização de
+ * uma página web.
+ * Como se trata de um fluxo assíncrono, é necessário que
+ * o usuário saiba que esses dados estão sendo carregados,
+ * para tanto você deve ativar uma componente de
+ * "Carregando" através da função setLoading.
  *
- * 1. O MDC de um número N com 0 é o próprio N.
+ * Para usar essa função basta chamar em seu código:
+ * setLoading(true), para abrir essa componente e
+ * setLoading(false), para fechá-la.
  *
- * 2. O MDC entre dois números M e N, onde M > N é
- * igual ao MDC de N e do resto da divisão de M por N.
+ * Uma vez que você obtiver os dados, você deve armazená-
+ * los através de uma função setData.
  *
- * Você pode considerar que nas entradas dos testes
- * a > b sempre.
+ * Em caso de erro, você deve abrir uma componente de
+ * erro, através da função setError, que funciona analoga-
+ * mente à função de loading.
+ *
+ * Para fazer a request, use nossa função chamada api, na
+ * forma api().then()...
+ *
  */
 
-const MDC = (a, b) => {}
+const { setData, setError, setLoading } = require('../../utils/stateHandlers')
+const api = require('../../utils/api') // Promise
 
-module.exports = MDC
+const doRequest = () => {}
+
+module.exports = doRequest
