+++ challenges/6/challenge.js
@@ -1,31 +1,51 @@
 /*
- * Regex
+ * Soma Combinacional
  */
-
 /* ENUNCIADO
- *
- * Escreva uma função que busque no texto os valores de "height" e "width"
- * e retorne um objeto { "height": x, "width": y } contendo esses valores ignorando sua medida (px, %, em).
- *
- * Ex:1
- * [INPUT]
- * "<div style="height: 20px; width: 60px;">"
- * [OUTPUT]
- * {
- *   height: 20,
- *   width: 60
- * }
- *
- * Ex: 2
- * [INPUT] String
- * "<div style="background-color: red;"> <img style="width: 120px; height: 20%" /></div>"
- * [OUTPUT] Object
- * {
- *   width: 120,
- *   height: 20
- * }
+ * Dado um conjunto de dados e um número alvo, você deve encontrar todas as combinações exclusivas 
+ * entre valores do conjunto que resultem na soma do número alvo.
+ * 
+ * Observações:
+ * 
+ * Todos os números, inclusive o alvo, serão inteiros positivos
+ * O resultado não deve obter combinações com valores repetidos. Exemplo:
+ *  
+ *  combinate([2, 3, 5], 8) retornando
+ * 
+ *  [
+ *       [2,2,2,2],
+ *       [2,3,3],
+ *       [3,2,3],
+ *       [3,3,2]  
+ *       [3,5]
+ *       [5,3]
+ *   ]
+ * 
+ * Os valores do conjunto de dados podem se repetir entre si, como por exemplo:
+ * 
+ * combinate([2, 3, 5], 8) retornando [2,2,2,2] e [2,3,3] como conjuntos que resultam na soma alvo.
+ * 
+ * 
+ * Seguem mais alguns exemplos do retorno esperado:
+ * 
+ *  combinate([2, 3, 5], 8) deve retornar
+ * 
+ *  [
+ *       [2,2,2,2],
+ *       [2,3,3],
+ *       [3,5]
+ *   ]
+*
+*    outro exemplo: 
+*
+*    combinate([2, 3, 6, 7], 7) retorna
+*
+*    [
+*       [2, 2, 3],
+*       [7]
+*    ]
  */
 
-const extractSize = htmlTemplate => {}
+const combinate = (set, target) => {}
 
-module.exports = extractSize
+module.exports = combinate
