--- utils/isPrime.js
@@ -1,6 +1,0 @@
-const isPrime = num => {
-  for (let i = 2; i < num; i++) if (num % i === 0) return false
-  return num > 1
-}
-
-module.exports = isPrime
